fix: revert Prometheus types from foundation back to model-serving/shared
Foundation is for generic language-level utilities, not domain-specific
API wire types. Prometheus query-range types belong in model-serving/shared
since model-serving is the feature owner for metrics. Consumers like gpuaas
use import type so no runtime edge is introduced.

Also moves PrometheusQueryRangeResponseData and PrometheusQueryRangeResponse
out of frontend/src/types.ts into model-serving/shared/types.ts to
consolidate all Prometheus range types in one canonical location.
