+++ frontend/src/__mocks__/mockPrometheusBias.ts
@@ -2,7 +2,7 @@
 import type {
   PrometheusQueryRangeResponse,
   PrometheusQueryRangeResponseDataResult,
-} from '@odh-dashboard/foundation';
+} from '@odh-dashboard/model-serving/shared/types';
 
 type MockPrometheusServingType = {
   result?: PrometheusQueryRangeResponseDataResult[];
