+++ frontend/src/api/prometheus/useQueryRangeResourceData.ts
@@ -2,7 +2,7 @@ import React from 'react';
 import { FetchOptions } from '@odh-dashboard/ui-core/hooks/useFetchState';
 import { TimeframeTitle } from '@odh-dashboard/ui-core/types/metrics';
 import { TimeframeStep, TimeframeTimeRange } from '@odh-dashboard/ui-core/utilities/metrics';
-import type { PrometheusQueryRangeResultValue } from '@odh-dashboard/foundation';
+import type { PrometheusQueryRangeResultValue } from '@odh-dashboard/model-serving/shared/types';
 import { PendingContextResourceData } from '#~/types';
 import { useMakeFetchObject } from '#~/utilities/useMakeFetchObject';
 import usePrometheusQueryRange, { ResponsePredicate } from './usePrometheusQueryRange';
