+++ packages/cypress/package.json
@@ -27,12 +27,12 @@
     "@odh-dashboard/app-config": "*",
     "@odh-dashboard/eslint-config": "*",
     "@odh-dashboard/feature-store": "*",
-    "@odh-dashboard/foundation": "*",
     "@odh-dashboard/hardware-profiles": "*",
     "@odh-dashboard/internal": "*",
     "@odh-dashboard/k8s-core": "*",
     "@odh-dashboard/maas": "*",
     "@odh-dashboard/model-registry": "*",
+    "@odh-dashboard/model-serving": "*",
     "@odh-dashboard/model-training": "*",
     "@odh-dashboard/plugin-core": "*",
     "@odh-dashboard/ui-core": "*",
