+++ packages/gpuaas/src/hooks/useBorrowingLendingMetrics.ts
@@ -5,7 +5,7 @@ import type {
   PrometheusQueryRangeResponseData,
   PrometheusQueryRangeResponseDataResult,
   PrometheusQueryRangeResultValue,
-} from '@odh-dashboard/foundation';
+} from '@odh-dashboard/model-serving/shared/types';
 import {
   ACCELERATOR_RESOURCE_PREFIXES,
   ACCELERATOR_RESOURCE_REGEX,
