+++ packages/model-serving/src/components/metrics/ModelServingMetricsContext.tsx
@@ -1,13 +1,13 @@
 import * as React from 'react';
-import type {
-  PrometheusQueryRangeResponseDataResult,
-  PrometheusQueryRangeResultValue,
-} from '@odh-dashboard/foundation';
 import { FetchStateObject } from '@odh-dashboard/ui-core/hooks/useFetch';
 import { DEFAULT_LIST_FETCH_STATE } from '@odh-dashboard/ui-core/utilities/fetchState';
 import { useModelServingMetrics } from '@odh-dashboard/internal/api';
 import { MetricsCommonContext } from '@odh-dashboard/internal/concepts/metrics/MetricsCommonContext';
-import { PerformanceMetricType } from '@odh-dashboard/model-serving/shared/types';
+import {
+  PerformanceMetricType,
+  PrometheusQueryRangeResponseDataResult,
+  PrometheusQueryRangeResultValue,
+} from '@odh-dashboard/model-serving/shared/types';
 
 export enum ModelMetricType {
   REQUEST_COUNT_SUCCESS = 'inference_request-count-successes',
